+++ src/tui/models/mod.rs
@@ -1,2 +1,2 @@
-pub mod app_state;
-pub mod tree_node;
+pub mod node;
+pub mod state;
