+++ src/tui/models/node.rs
